requirements.txt
numpy>=1.12.1
pillow>=4.0.0
pydicom>=0.9.9